[REFACTOR] #9: 폴더 구조 변경
