+++ src/main/java/SeoulMilk1_BE/post/dto/request/comment/CommentCreateRequest.java
@@ -1,4 +1,4 @@
-package SeoulMilk1_BE.post.dto.request;
+package SeoulMilk1_BE.post.dto.request.comment;
 
 public record CommentCreateRequest(Long postId, Long userId, String text) {
 }
