+++ src/main/java/SeoulMilk1_BE/post/dto/request/comment/CommentUpdateRequest.java
@@ -1,4 +1,4 @@
-package SeoulMilk1_BE.post.dto.request;
+package SeoulMilk1_BE.post.dto.request.comment;
 
 public record CommentUpdateRequest(String text) {
 }
