+++ src/main/java/SeoulMilk1_BE/post/dto/request/post/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package SeoulMilk1_BE.post.dto.request;
+package SeoulMilk1_BE.post.dto.request.post;
 
 public record PostUpdateRequest(String title, String content, String type) {
 }
