+++ src/main/java/SeoulMilk1_BE/post/dto/response/comment/CommentReadResponse.java
@@ -1,4 +1,4 @@
-package SeoulMilk1_BE.post.dto.response;
+package SeoulMilk1_BE.post.dto.response.comment;
 
 public record CommentReadResponse(Long commentId, String username, String text) {
 }
