+++ src/main/java/SeoulMilk1_BE/post/dto/response/comment/CommentUpdateResponse.java
@@ -1,4 +1,4 @@
-package SeoulMilk1_BE.post.dto.response;
+package SeoulMilk1_BE.post.dto.response.comment;
 
 import java.time.LocalDateTime;
